Add listen to article functions to football template
